% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea-plot.R
\name{gsea_plot}
\alias{gsea_plot}
\title{Plot GSEA Enrichment Score}
\usage{
gsea_plot(gsea_res, set_names = gsea_res[["pathway"]], show_table = FALSE)
}
\arguments{
\item{gsea_res}{A result object from \code{run_microbiome_gsea}. It should be a
data frame with "sampleSets" and "sampleRanks" attributes.}

\item{set_names}{A character vector The name of the sets (pathway) to plot.}

\item{show_table}{A logical value. If \code{TRUE}, shows a table of p-values and NES alongside the plot.}
}
\value{
A ggplot object (generated by \code{fgsea::plotEnrichment}).
}
\description{
This function plots the GSEA enrichment score curve for a specific sample set (pathway)
using the results from \code{run_microbiome_gsea}.
}
\examples{
\dontrun{
gsea_res = run_microbiome_gsea(ps_test_data, "Genus1", "Genus", c("Group","Source"))
gsea_plot(gsea_res, "Group_Treat")
}
}
