% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea-plot.R
\name{gsea_plot_barcode}
\alias{gsea_plot_barcode}
\title{Plot GSEA Enrichment Score Barcodes for Multiple Sets}
\usage{
gsea_plot_barcode(gsea_res, set_names = gsea_res[["pathway"]])
}
\arguments{
\item{gsea_res}{A result object from \code{run_microbiome_gsea}. It should be a
data frame with "sampleSets" and "sampleRanks" attributes.}

\item{set_names}{A character vector The names of the sets (pathways) to plot.}
}
\value{
A ggplot object (combined barcode plots for the specified pathways).
}
\description{
This function generates barcode plots for multiple GSEA pathways, displaying the enrichment score ticks
along with pathway names and statistics (size, NES, adjusted p-value).
}
