% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea-run.R
\name{run_microbiome_gsea}
\alias{run_microbiome_gsea}
\title{执行微生物群落样本级富集分析 (Sample-Level Enrichment Analysis)}
\usage{
run_microbiome_gsea(
  ps,
  taxon_target,
  taxon_rank,
  sample_feat,
  min_size = 5,
  max_size = 500,
  descreasing = TRUE
)
}
\arguments{
\item{ps}{A \code{phyloseq} object.}

\item{taxon_target}{A character string. The target taxon name (e.g., "Escherichia").}

\item{taxon_rank}{A character string. The rank of the target taxon (e.g., "Genus").}

\item{sample_feat}{A character string. The metadata column name to group samples (e.g., "Group").}

\item{min_size}{Integer. Minimum size of a feature set. Default is 5.}

\item{max_size}{Integer. Maximum size of a feature set. Default is 500.}

\item{descreasing}{Logical. Whether to sort in decreasing order. Default is TRUE.}
}
\value{
A \code{data.frame} (tbl_df) containing the GSEA results.
Columns include: pathway, pval, padj, NES, size, leadingEdge.
}
\description{
该函数接受 phyloseq 对象，针对特定的目标微生物，分析样本元数据特征是否在
该微生物高丰度或低丰度的样本中富集。
}
\examples{
\dontrun{
# Suppose there is a ps object.
res <- run_microbiome_gsea(
    ps = ps_test_data,
    taxon_target = "Genus1",
    taxon_rank = "Genus",
    sample_feat = c("Group", "Source")
)
print(res)
}
}
